+++ converter/zip/zip_test.go
@@ -6,6 +6,7 @@ import (
 	"github.com/metafates/mangal/config"
 	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/filesystem"
+	"github.com/metafates/mangal/key"
 	"github.com/metafates/mangal/source"
 	"github.com/samber/lo"
 	. "github.com/smartystreets/goconvey/convey"
@@ -18,7 +19,7 @@ import (
 func init() {
 	filesystem.SetMemMapFs()
 	lo.Must0(config.Setup())
-	viper.Set(constant.FormatsUse, constant.FormatZIP)
+	viper.Set(key.FormatsUse, constant.FormatZIP)
 }
 
 func TestCBZ(t *testing.T) {