+++ downloader/download.go
@@ -4,10 +4,10 @@ import (
 	"encoding/json"
 	"fmt"
 	"github.com/metafates/mangal/color"
-	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/converter"
 	"github.com/metafates/mangal/filesystem"
 	"github.com/metafates/mangal/history"
+	"github.com/metafates/mangal/key"
 	"github.com/metafates/mangal/log"
 	"github.com/metafates/mangal/source"
 	"github.com/metafates/mangal/style"
@@ -25,7 +25,7 @@ func Download(chapter *source.Chapter, progress func(string)) (string, error) {
 		return "", err
 	}
 
-	if viper.GetBool(constant.DownloaderRedownloadExisting) {
+	if viper.GetBool(key.DownloaderRedownloadExisting) {
 		log.Info("chapter already downloaded, deleting and redownloading")
 		err = filesystem.Api().Remove(path)
 		if err != nil {
@@ -53,14 +53,14 @@ func Download(chapter *source.Chapter, progress func(string)) (string, error) {
 		return "", err
 	}
 
-	if viper.GetBool(constant.MetadataFetchAnilist) {
+	if viper.GetBool(key.MetadataFetchAnilist) {
 		err := chapter.Manga.PopulateMetadata(progress)
 		if err != nil {
 			log.Warn(err)
 		}
 	}
 
-	if viper.GetBool(constant.MetadataSeriesJSON) {
+	if viper.GetBool(key.MetadataSeriesJSON) {
 		path, err := chapter.Manga.Path(false)
 		if err != nil {
 			log.Warn(err)
@@ -80,34 +80,34 @@ func Download(chapter *source.Chapter, progress func(string)) (string, error) {
 		}
 	}
 
-	if viper.GetBool(constant.DownloaderDownloadCover) {
+	if viper.GetBool(key.DownloaderDownloadCover) {
 		coverDir, err := chapter.Manga.Path(false)
 		if err == nil {
 			_ = chapter.Manga.DownloadCover(false, coverDir, progress)
 		}
 	}
 
-	log.Info("getting " + viper.GetString(constant.FormatsUse) + " converter")
+	log.Info("getting " + viper.GetString(key.FormatsUse) + " converter")
 	progress(fmt.Sprintf(
 		"Converting %d pages to %s %s",
 		len(pages),
-		style.Fg(color.Yellow)(viper.GetString(constant.FormatsUse)),
+		style.Fg(color.Yellow)(viper.GetString(key.FormatsUse)),
 		style.Faint(chapter.SizeHuman())),
 	)
-	conv, err := converter.Get(viper.GetString(constant.FormatsUse))
+	conv, err := converter.Get(viper.GetString(key.FormatsUse))
 	if err != nil {
 		log.Error(err)
 		return "", err
 	}
 
-	log.Info("converting " + viper.GetString(constant.FormatsUse))
+	log.Info("converting " + viper.GetString(key.FormatsUse))
 	path, err = conv.Save(chapter)
 	if err != nil {
 		log.Error(err)
 		return "", err
 	}
 
-	if viper.GetBool(constant.HistorySaveOnDownload) {
+	if viper.GetBool(key.HistorySaveOnDownload) {
 		go func() {
 			err = history.Save(chapter)
 			if err != nil {