+++ key/keys.go
@@ -1,9 +1,9 @@
-package constant
+package key
 
 // DefinedFieldsCount is the number of fields defined in this package.
 // You have to manually update this number when you add a new field
 // to check later if every field has a defined default value
-const DefinedFieldsCount = 51
+const DefinedFieldsCount = 52
 
 const (
 	DownloaderPath                = "downloader.path"
@@ -96,4 +96,8 @@ const (
 	LogsWrite = "logs.write"
 	LogsLevel = "logs.level"
 	LogsJson  = "logs.json"
+)
+
+const (
+	CliColored = "cli.colored"
 )