+++ provider/mangadex/search.go
@@ -3,7 +3,7 @@ package mangadex
 import (
 	"fmt"
 	"github.com/darylhjd/mangodex"
-	"github.com/metafates/mangal/constant"
+	"github.com/metafates/mangal/key"
 	"github.com/metafates/mangal/source"
 	"github.com/spf13/viper"
 	"log"
@@ -29,7 +29,7 @@ func (m *Mangadex) Search(query string) ([]*source.Manga, error) {
 		params.Add("contentRating[]", rating)
 	}
 
-	if viper.GetBool(constant.MangadexNSFW) {
+	if viper.GetBool(key.MangadexNSFW) {
 		params.Add("contentRating[]", mangodex.Porn)
 		params.Add("contentRating[]", mangodex.Erotica)
 	}
@@ -47,7 +47,7 @@ func (m *Mangadex) Search(query string) ([]*source.Manga, error) {
 
 	for i, manga := range mangaList.Data {
 		m := source.Manga{
-			Name:   manga.GetTitle(viper.GetString(constant.MangadexLanguage)),
+			Name:   manga.GetTitle(viper.GetString(key.MangadexLanguage)),
 			URL:    fmt.Sprintf("https://mangadex.org/title/%s", manga.ID),
 			Index:  uint16(i),
 			ID:     manga.ID,