+++ tui/bubble.go
@@ -11,9 +11,9 @@ import (
 	"github.com/charmbracelet/lipgloss"
 	"github.com/metafates/mangal/anilist"
 	"github.com/metafates/mangal/color"
-	"github.com/metafates/mangal/constant"
 	"github.com/metafates/mangal/history"
 	"github.com/metafates/mangal/installer"
+	key2 "github.com/metafates/mangal/key"
 	"github.com/metafates/mangal/provider"
 	"github.com/metafates/mangal/source"
 	"github.com/metafates/mangal/style"
@@ -196,7 +196,7 @@ func newBubble() *statefulBubble {
 
 	makeList := func(title string, description bool, options *listOptions) list.Model {
 		delegate := list.NewDefaultDelegate()
-		delegate.SetSpacing(viper.GetInt(constant.TUIItemSpacing))
+		delegate.SetSpacing(viper.GetInt(key2.TUIItemSpacing))
 		delegate.ShowDescription = description
 		delegate.Styles.SelectedTitle = lipgloss.NewStyle().
 			Border(lipgloss.ThickBorder(), false, false, false, true).
@@ -234,7 +234,7 @@ func newBubble() *statefulBubble {
 	bubble.inputC = textinput.New()
 	bubble.inputC.Placeholder = "Search"
 	bubble.inputC.CharLimit = 60
-	bubble.inputC.Prompt = viper.GetString(constant.TUISearchPromptString)
+	bubble.inputC.Prompt = viper.GetString(key2.TUISearchPromptString)
 
 	bubble.progressC = progress.New(progress.WithDefaultGradient())
 
@@ -255,7 +255,7 @@ func newBubble() *statefulBubble {
 	})
 	bubble.sourcesC.SetStatusBarItemName("source", "sources")
 
-	showURLs := viper.GetBool(constant.TUIShowURLs)
+	showURLs := viper.GetBool(key2.TUIShowURLs)
 	bubble.mangasC = makeList("Mangas", showURLs, &listOptions{
 		TitleStyle: mo.Some(
 			style.NewColored("#f2e8cf", "#386641").Padding(0, 1),