+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to
 [Semantic Versioning](https://semver.org).
 
+## 4.0.5
+
+- Fixes runtime crash #135
+- Option to disable colors in cli help `mangal config info -k cli.colored` (why not? =P)
+- Improved `config info` command output. It now shows default value and env variable name.
+- Internal improvements
+
 ## 4.0.4
 
 - Fix manga tags and genres being the same inside ComicInfo.xml #133